Construct Displacement Sphere

Construct Displacement Sphere

displacement_sphere(
  displacement_texture,
  displacement_scale = 1,
  use_cube = FALSE,
  cube_subdivision_levels = NA,
  displace = TRUE,
  verbose = TRUE,
  position = c(0, 0, 0),
  scale = c(1, 1, 1),
  angle = c(0, 0, 0),
  pivot_point = c(0, 0, 0),
  order_rotation = c(1, 2, 3),
  material = material_list()
)

Arguments

displacement_texture

Image or matrix/array that will be used to displace the sphere.

displacement_scale

Default 1. Scale of the displacement.

use_cube

Default FALSE. Whether to use a subdivided cube instead of a UV sphere. Use this if you want to visualize areas near the poles.

cube_subdivision_levels

Default NA. Uses the dimensions of the displacement texture to automatically calculate the number of subdivision levels.

displace

Default TRUE. Whether to displace the sphere, or just generate the initial mesh for later displacement.

verbose

Default TRUE. Whether to print displacement texture information.

position

Default c(0,0,0). Position of the mesh.

scale

Default c(1,1,1). Scale of the mesh. Can also be a single numeric value scaling all axes uniformly.

angle

Default c(0,0,0). Angle to rotate the mesh.

pivot_point

Default c(0,0,0). Point around which to rotate the mesh.

order_rotation

Default c(1,2,3). Order to rotate the axes.

material

Default material_list() (default values). Specify the material of the object.

Value

raymesh object

Examples

if(run_documentation()) {

}
#> NULL